Move peer module to separate file
